separated batched protocol messages into a separate class
- This allows a user to call messages over any communication medium,
not just GAE.
- TODO: create direct message option, rather than batched messages.
- TODO: this code is still buggy even using batched messages.
