+++ safeslinger-exchange/src/edu/cmu/cylab/starslinger/exchange/ConnectionEngine.java
@@ -81,161 +81,69 @@ public abstract class ConnectionEngine {
     /**
      * send commitment, receives unique short user id
      */
-    public byte[] assign_user(byte[] commitB) throws ExchangeException {
-        ByteBuffer msg = ByteBuffer.allocate(4 + commitB.length);
-        msg.putInt(mVersion);
-        msg.put(commitB);
+    public byte[] assign_user(byte[] msg) throws ExchangeException {
 
-        byte[] resp = assignUser(msg.array());
+        byte[] resp = assignUser(msg);
         handleResponseExceptions(resp, 0);
 
         Log.i(TAG, "User id created");
         return resp;
     }
 
-    /**
-     * send our own matches once, list of all users we have, gathers all others
-     * when available receives the group id, total user number (including ours),
-     * actual user number, list of users we did not get yet
-     */
-    public byte[] sync_commits(int usrid, int sameusrid, int[] usridList, byte[] commitB)
-            throws ExchangeException {
+    public byte[] sync_commits(byte[] msg) throws ExchangeException {
 
         handleTimeoutException();
-
-        ByteBuffer msg = ByteBuffer.allocate(4 + 4 + 4 + 4 + (usridList.length * 4)
-                + commitB.length);
-        msg.putInt(mVersion);
-        msg.putInt(usrid);
-        msg.putInt(sameusrid);
-        msg.putInt(usridList.length);
-        for (int user : usridList) {
-            msg.putInt(user);
-        }
-        msg.put(commitB);
-
-        byte[] resp = syncUsers(msg.array());
+        byte[] resp = syncUsers(msg);
         handleResponseExceptions(resp, 0);
 
         Log.i(TAG, "User created");
         return resp;
     }
 
-    /**
-     * send our own data once, list of all data we have, gathers all others when
-     * available receives the total data number (including ours), actual data
-     * number, list of data we did not get yet
-     */
-    public byte[] sync_data(int usrid, int[] usridList, byte[] data) throws ExchangeException {
+    public byte[] sync_data(byte[] msg) throws ExchangeException {
 
         handleTimeoutException();
-
-        ByteBuffer msg = ByteBuffer.allocate(4 + 4 + 4 + (usridList.length * 4) + data.length);
-        msg.putInt(mVersion);
-        msg.putInt(usrid);
-        msg.putInt(usridList.length);
-        for (int user : usridList) {
-            msg.putInt(user);
-        }
-        msg.put(data);
-
-        byte[] resp = syncData(msg.array());
+        byte[] resp = syncData(msg);
         handleResponseExceptions(resp, 0);
 
         Log.i(TAG, "User updated");
         return resp;
     }
 
-    /**
-     * send our own signature once, list of all signatures we have, gathers all
-     * others when available receives the total signatures number (including
-     * ours), actual sig number, list of signatures we did not get yet
-     */
-    public byte[] sync_signatures(int usrid, int[] usridList, byte[] signature)
-            throws ExchangeException {
+    public byte[] sync_signatures(byte[] msg) throws ExchangeException {
 
         handleTimeoutException();
-
-        ByteBuffer msg = ByteBuffer.allocate(4 + 4 + 4 + (usridList.length * 4) + signature.length);
-        msg.putInt(mVersion);
-        msg.putInt(usrid);
-        msg.putInt(usridList.length);
-        for (int user : usridList) {
-            msg.putInt(user);
-        }
-        msg.put(signature);
-
-        byte[] resp = syncSignatures(msg.array());
+        byte[] resp = syncSignatures(msg);
         handleResponseExceptions(resp, 0);
 
         Log.i(TAG, "Signature sent");
         return resp;
     }
 
-    /**
-     * this method is used by members to post one public key node. send: node to
-     * submit (user id, node length, node);
-     */
-    public byte[] put_keynode(int usrid, int nodeusrid, byte[] node) throws ExchangeException {
+    public byte[] put_keynode(byte[] msg) throws ExchangeException {
 
         handleTimeoutException();
-
-        ByteBuffer msg = ByteBuffer.allocate(4 + 4 + 4 + 4 + node.length);
-        msg.putInt(mVersion);
-        msg.putInt(usrid);
-        msg.putInt(nodeusrid);
-        msg.putInt(node.length);
-        msg.put(node);
-
-        byte[] resp = syncKeyNodes(msg.array());
+        byte[] resp = syncKeyNodes(msg);
         handleResponseExceptions(resp, 2);
 
         Log.i(TAG, "key node sent");
         return resp;
     }
 
-    /**
-     * this method is used by members to discover if their key node node has
-     * been submitted. receive: the total nodes number for themselves (0 or 1),
-     * our own key node if available.
-     */
-    public byte[] get_keynode(int usrid) throws ExchangeException {
+    public byte[] get_keynode(byte[] msg) throws ExchangeException {
 
         handleTimeoutException();
-
-        ByteBuffer msg = ByteBuffer.allocate(4 + 4);
-        msg.putInt(mVersion);
-        msg.putInt(usrid);
-
-        byte[] resp = syncKeyNodes(msg.array());
+        byte[] resp = syncKeyNodes(msg);
         handleResponseExceptions(resp, 2);
 
         Log.i(TAG, "key node requested");
         return resp;
     }
 
-    /**
-     * send our own match nonce once, list of all match nonces we have, gathers
-     * all others when available receives the total match nonces number
-     * (including ours), actual match nonce number, list of match nonces we did
-     * not get yet
-     */
-    public byte[] sync_match(int usrid, int[] usridList, byte[] matchNonce)
-            throws ExchangeException {
+    public byte[] sync_match(byte[] msg) throws ExchangeException {
 
         handleTimeoutException();
-
-        ByteBuffer msg = ByteBuffer
-                .allocate(4 + 4 + 4 + (usridList.length * 4) + matchNonce.length);
-        msg.putInt(mVersion);
-        msg.putInt(usrid);
-        msg.putInt(usridList.length);
-        for (int user : usridList) {
-            msg.putInt(user);
-        }
-        msg.put(matchNonce);
-
-        byte[] resp = syncMatch(msg.array());
+        byte[] resp = syncMatch(msg);
         handleResponseExceptions(resp, 0);
 
         Log.i(TAG, "Match nonce sent");
